+++ clients/vscode-hlasmplugin/CHANGELOG.md
@@ -17,6 +17,7 @@
 - The language server may crash when a complex expression is used as a variable symbol index
 - References to CA variables in strings are not reported
 - Structured macro variables were not forwarded correctly when a dot separator was used in the macro operand
+- Return correct variable type for values provided in the macro's name field
 
 ## [1.2.0](https://github.com/eclipse/che-che4z-lsp-for-hlasm/compare/1.1.0...1.2.0) (2022-05-11)